smarttrade-func/requirements.txt
azure-functions==1.21.3
azure-storage-queue==12.11.0
azure-data-tables==12.5.0
requests==2.32.3
